requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.3.2
fastapi==0.104.1
idna==3.4
pydantic==2.4.2
pydantic_core==2.10.1
python-restcountries==2.0.0
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.7
